vistaone-api/requirements.txt
alembic==1.18.4
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.1
colorama==0.4.6
Deprecated==1.3.1
ecdsa==0.19.1
Flask==3.1.3
flask-cors==6.0.2
Flask-Limiter==4.1.1
Flask-Mail==0.10.0
flask-marshmallow==1.4.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
flask-swagger==0.2.14
flask-swagger-ui==5.21.0
greenlet==3.3.2
gunicorn==25.3.0
idna==3.10
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.8.0
Mako==1.3.10
MarkupSafe==3.0.3
marshmallow==4.2.2
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==1.4.2
mysql-connector-python==9.5.0
ordered-set==4.1.0
packaging==26.0
pluggy==1.6.0
psycopg2==2.9.11
psycopg2-binary==2.9.11
pyasn1==0.6.3
Pygments==2.19.2
PyJWT==2.12.1
pytest==9.0.2
python-dotenv==1.2.2
python-jose==3.5.0
PyYAML==6.0.3
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.48
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.5.0
Werkzeug==3.1.6
wrapt==2.1.2
